Adicionando algumas verificações no service de users

+++ src/user/user.service.ts
@@ -1,13 +1,13 @@
 import { BadRequestException, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Users } from './entity/users.entity';
-import { GetUsersDto } from './dto/getUsers.dto';
+import { Users } from './user.entity';
+import { GetUserDto } from './dto/getUser.dto';
 import * as bcrypt from 'bcrypt';
-import { UpdateUsersDto } from './dto/updateUsers.dto';
+import { UpdateUserDto } from './dto/updateUser.dto';
 
 @Injectable()
-export class UsersService {
+export class UserService {
   constructor(
     @InjectRepository(Users)
     private readonly userRepository: Repository<Users>,
@@ -19,13 +19,19 @@ export class UsersService {
     return hash;
   }
 
-  async findAll(): Promise<GetUsersDto[]> {
+  async findAll(): Promise<GetUserDto[]> {
     return await this.userRepository.find({
       select: ['id', 'firstName', 'lastName', 'email'],
     });
   }
 
-  async findOne(id: number): Promise<GetUsersDto> {
+  async findOne(id: number): Promise<GetUserDto> {
+    const hasUser = await this.userRepository.findOne({ where: { id } });
+
+    if (!hasUser) {
+      throw new BadRequestException(`O usuário: ${id}, não existe`);
+    }
+
     try {
       return await this.userRepository.findOne({
         where: { id },
@@ -39,6 +45,14 @@ export class UsersService {
   }
 
   async create(user: Users): Promise<Users> {
+    const hasEmail = await this.userRepository.findOne({
+      where: { email: user.email },
+    });
+
+    if (hasEmail) {
+      throw new BadRequestException('O e-mail informado já foi cadastrado');
+    }
+
     try {
       const hashedPassword = await this.hashPassword(user.password);
       return await this.userRepository.save(
@@ -50,21 +64,31 @@ export class UsersService {
     } catch (error) {
       console.log(error);
 
-      throw new BadRequestException('Não foi possível criar usuário!');
+      throw new BadRequestException('Não foi possível criar usuário');
     }
   }
 
-  async update(id: number, user: UpdateUsersDto): Promise<Users> {
-    console.log({ user });
+  async update(id: number, user: UpdateUserDto): Promise<Users> {
+    try {
+      await this.userRepository.update(id, user);
+      return await this.userRepository.findOne({
+        where: { id },
+        select: ['id', 'firstName', 'lastName', 'email'],
+      });
+    } catch (error) {
+      console.log(error);
 
-    await this.userRepository.update(id, user);
-    return await this.userRepository.findOne({
-      where: { id },
-      select: ['id', 'firstName', 'lastName', 'email'],
-    });
+      throw new BadRequestException(`Não foi possível editar o usuário: ${id}`);
+    }
   }
 
   async delete(id: number): Promise<void> {
+    const hasUser = await this.userRepository.findOne({ where: { id } });
+
+    if (!hasUser) {
+      throw new BadRequestException(`O usuário: ${id}, não existe`);
+    }
+
     try {
       await this.userRepository.delete(id);
     } catch (error) {
